indexador_documentos/requirements.txt
pymupdf
pytesseract
Pillow
pywebview
# OCR Transformer opcional (pesado):
transformers
torch
gradio
